mod: UI 수정 및 컴포넌트화
